optical-flow-gui
This GUI offers two display options, both describe the nature of optical flow data being processed by the Avoidance-Drone.

The Left-Right-Balancing option displays a conjoined graphs.  The Red Line displays the magnitude of the right screen’s flow.  The Blue Line displays the magnitude of the left screen’s flow.  The Purple Line displays the difference between the two and indicates which way the drone will turn in response.

The Focal-Point option displays the optical flow data of the inner pixels and when the flow threshold has been, at least, met.